refactor: renamed channel font to badge font
